backend/app/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.6
fonttools==4.55.3
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
zipp==3.21.0
